% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_EZ.R
\name{format_databars}
\alias{format_databars}
\title{Add Databars to the Datatable}
\usage{
format_databars(table, columns, color = "Pastel1", min_quantile = 0.1)
}
\arguments{
\item{table}{an object from \code{DT::datatable} or
\code{finTrendr::datatable_EZ} output}

\item{columns}{columns to be formatted (can be character or a formula of the
form \code{~ V1 + V2}, which is equivalent to \code{c('V1', 'V2')})}

\item{color}{Background colors for each column (recycled if needed). One of:
\itemize{
  \item One of \code{RColorBrewer::brewer.pal()} qualitative palettes
    ('Accent', 'Dark2','Paired', 'Pastel1', 'Pastel2', 'Set1', 'Set2', 'Set3'),
    or a numeric index 1-8 of those names. The default is "Pastel1"
  \item valid R color name(s) - like \code{"red3"}
  \item RGB value(s) - like \code{"rgb(255, 0, 0)"}
  \item HEX value(s) - like \code{"#ff0000"}
}}

\item{min_quantile}{Controls the minimum length of the databar. The row with
the maximum value in a column will always have a full bar. The row with the
minimum value will have a zero-length bar, if \code{min_quantile=0}.
Default is 0.10.}
}
\value{
Formatted \code{DT::datatable} object
}
\description{
Applies conditional formatting databars to one or more columns in a
\code{DT::datatable()}.
}
\examples{
dt <- iris \%>\% datatable_EZ()

# multiple columns specification
# same as c("Sepal.Length","Petal.Length")
dt \%>\% format_databars(~ Sepal.Length + Petal.Length)

# Specify R color names
dt \%>\% format_databars(~ Sepal.Length, color=c("red3","grey75"))

# Specify RColorBrewer qualitative palettes
dt \%>\% format_databars(~ Sepal.Length, color= "Pastel1") # by name
dt \%>\% format_databars(~ Sepal.Length, color= 2) # by index

}
\seealso{
\code{\link{datatable_EZ}}, \code{\link[DT]{datatable}},
  \code{\link[DT]{formatStyle}}, \code{\link[DT]{styleColorBar}}
}
