% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_EZ.R
\name{format_databars}
\alias{format_databars}
\title{Add Databars to the Datatable}
\usage{
format_databars(table, columns, color = "Pastel1", min_quantile = 0.1)
}
\arguments{
\item{table}{an object from \code{DT::datatable} or
\code{finTrendr::datatable_EZ} output}

\item{columns}{columns to be formatted (can be character or a formula of the
form \code{~ V1 + V2}, which is equivalent to \code{c('V1', 'V2')})}

\item{color}{a character vector of color names (will be recycled).
Alternatively, a character name of one of the
\code{RColorBrewer::brewer.pal()} qualitative palettes ('Accent', 'Dark2',
'Paired', 'Pastel1', 'Pastel2', 'Set1', 'Set2', 'Set3'), or a numeric index
1-8 of those names. The default is "Pastel1".}

\item{min_quantile}{Controls the minimum length of the databar. The row with
the maximum value in a column will always have a full bar. The row with the
minimum value will have a zero-length bar, if \code{min_quantile=0}.
Default is 0.10.}
}
\value{
Formatted \code{DT::datatable} object
}
\description{
Applies conditional formating databars to one or more columns in a
\code{DT::datatable()}.
}
\examples{
library(DT)
library(magrittr)

iris \%>\%
  datatable_EZ() \%>\%
  format_databars(~ Sepal.Length + Petal.Length) # same as c("Sepal.Length","Petal.Length")

}
\seealso{
\code{\link{datatable_EZ}}, \code{\link[DT]{datatable}},
  \code{\link[DT]{formatStyle}}, \code{\link[DT]{styleColorBar}}
}
